# More efficient parsing of ASCII strings for GraalVM EE