--- a/jsoniter-scala-core/src/main/scala/com/github/plokhotnyuk/jsoniter_scala/core/JsonReader.scala
+++ b/jsoniter-scala-core/src/main/scala/com/github/plokhotnyuk/jsoniter_scala/core/JsonReader.scala
@@ -2350,14 +2350,40 @@ final class JsonReader private[jsoniter_scala](
 
   @tailrec
   private[this] def parseString(i: Int, minLim: Int, charBuf: Array[Char], pos: Int): Int =
-    if (i < minLim) {
+    if (i + 3 < minLim) {
+      val b1 = buf(pos)
+      charBuf(i) = b1.toChar
+      val b2 = buf(pos + 1)
+      charBuf(i + 1) = b2.toChar
+      val b3 = buf(pos + 2)
+      charBuf(i + 2) = b3.toChar
+      val b4 = buf(pos + 3)
+      charBuf(i + 3) = b4.toChar
+      if (b1 == '"') {
+        head = pos + 1
+        i
+      } else if (((b1 - 32) ^ 60) <= 0) parseEncodedString(i, charBuf.length - 1, charBuf, pos)
+      else if (b2 == '"') {
+        head = pos + 2
+        i + 1
+      } else if (((b2 - 32) ^ 60) <= 0) parseEncodedString(i + 1, charBuf.length - 1, charBuf, pos + 1)
+      else if (b3 == '"') {
+        head = pos + 3
+        i + 2
+      } else if (((b3 - 32) ^ 60) <= 0) parseEncodedString(i + 2, charBuf.length - 1, charBuf, pos + 2)
+      else if (b4 == '"') {
+        head = pos + 4
+        i + 3
+      } else if (((b4 - 32) ^ 60) <= 0) parseEncodedString(i + 3, charBuf.length - 1, charBuf, pos + 3)
+      else parseString(i + 4, minLim, charBuf, pos + 4)
+    } else if (i < minLim) {
       val b = buf(pos)
       charBuf(i) = b.toChar
       if (b == '"') {
         head = pos + 1
         i
-      } else if (((b - 32) ^ 60) > 0) parseString(i + 1, minLim, charBuf, pos + 1) // == else if (b >= ' ' && b != '\\') ...
-      else parseEncodedString(i, charBuf.length - 1, charBuf, pos)
+      } else if (((b - 32) ^ 60) <= 0) parseEncodedString(i, charBuf.length - 1, charBuf, pos)
+      else parseString(i + 1, minLim, charBuf, pos + 1)
     } else if (pos >= tail) {
       val newPos = loadMoreOrError(pos)
       parseString(i, Math.min(charBuf.length, i + tail - newPos), charBuf, newPos)