Bellow are results of [Borer's benchmarks](https://github.com/sirthias/borer/tree/master/benchmarks) for jsoniter-scala.

# Before
### OpenJDK 11.0.3
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                                       (fileName)   Mode  Cnt      Score      Error  Units
[info] JsoniterScalaModelBenchmark.decodeModel                 australia-abc.json  thrpt    5  16705.498 ±  494.067  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                       bitcoin.json  thrpt    5  30533.002 ±  113.159  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                      doj-blog.json  thrpt    5   5057.857 ±  251.921  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel              eu-lobby-country.json  thrpt    5  38514.438 ± 1064.762  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel            eu-lobby-financial.json  thrpt    5   6363.730 ±  260.168  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                 eu-lobby-repr.json  thrpt    5   2461.462 ± 2042.136  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                 github-events.json  thrpt    5   3464.003 ±   82.481  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  github-gists.json  thrpt    5   8318.496 ±  257.043  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                json-generator.json  thrpt    5  39018.561 ±  303.682  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                    meteorites.json  thrpt    5   1381.890 ±   18.914  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                        movies.json  thrpt    5     96.922 ±    0.985  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  reddit-scala.json  thrpt    5   2337.139 ±   27.268  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                    rick-morty.json  thrpt    5  25193.538 ±   96.794  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  temp-anomaly.json  thrpt    5  45582.218 ±  580.777  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  thai-cinemas.json  thrpt    5  35834.050 ±   71.275  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                       turkish.json  thrpt    5    556.408 ±    2.380  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel  twitter_api_compact_response.json  thrpt    5  47258.646 ±  141.746  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel          twitter_api_response.json  thrpt    5  36598.778 ±  235.329  ops/s
```
### GraalVM CE 19.1.0
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                                       (fileName)   Mode  Cnt      Score      Error  Units
[info] JsoniterScalaModelBenchmark.decodeModel                 australia-abc.json  thrpt    5  18219.647 ±   49.762  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                       bitcoin.json  thrpt    5  27994.859 ±  142.715  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                      doj-blog.json  thrpt    5   3316.496 ±   99.059  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel              eu-lobby-country.json  thrpt    5  45226.505 ±  363.823  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel            eu-lobby-financial.json  thrpt    5   6437.307 ±  348.189  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                 eu-lobby-repr.json  thrpt    5   2418.977 ±  102.525  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                 github-events.json  thrpt    5   3430.517 ±  213.605  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  github-gists.json  thrpt    5   8282.431 ±  191.806  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                json-generator.json  thrpt    5  37686.127 ±  386.162  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                    meteorites.json  thrpt    5   1358.249 ±   14.376  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                        movies.json  thrpt    5     90.005 ±    2.131  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  reddit-scala.json  thrpt    5   3645.848 ±   21.999  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                    rick-morty.json  thrpt    5  26677.742 ±  325.127  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  temp-anomaly.json  thrpt    5  43536.346 ± 1464.688  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  thai-cinemas.json  thrpt    5  39323.691 ±  608.646  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                       turkish.json  thrpt    5    543.531 ±    3.080  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel  twitter_api_compact_response.json  thrpt    5  45190.303 ±  227.234  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel          twitter_api_response.json  thrpt    5  33051.229 ±  820.846  ops/s
```
### GraalVM EE 19.1.0
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                                       (fileName)   Mode  Cnt      Score      Error  Units
[info] JsoniterScalaModelBenchmark.decodeModel                 australia-abc.json  thrpt    5  16553.840 ±  207.368  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                       bitcoin.json  thrpt    5  25837.783 ±  933.601  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                      doj-blog.json  thrpt    5   3563.671 ±   69.878  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel              eu-lobby-country.json  thrpt    5  38422.096 ± 1883.551  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel            eu-lobby-financial.json  thrpt    5   6864.288 ±   86.696  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                 eu-lobby-repr.json  thrpt    5   2616.523 ±   24.593  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                 github-events.json  thrpt    5   3595.366 ±   22.156  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  github-gists.json  thrpt    5   7551.505 ±   67.057  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                json-generator.json  thrpt    5  34420.045 ±  262.423  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                    meteorites.json  thrpt    5   1310.555 ±   30.856  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                        movies.json  thrpt    5     92.512 ±    1.413  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  reddit-scala.json  thrpt    5   4148.875 ±   44.612  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                    rick-morty.json  thrpt    5  24506.937 ±  139.128  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  temp-anomaly.json  thrpt    5  44917.583 ±  802.078  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  thai-cinemas.json  thrpt    5  36134.004 ±  341.427  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                       turkish.json  thrpt    5    642.208 ±    2.758  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel  twitter_api_compact_response.json  thrpt    5  43155.412 ±  123.147  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel          twitter_api_response.json  thrpt    5  32023.644 ±  123.202  ops/s
```
# After
### OpenJDK 11.0.3
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                                       (fileName)   Mode  Cnt      Score      Error  Units
[info] JsoniterScalaModelBenchmark.decodeModel                 australia-abc.json  thrpt    5  18331.154 ± 1952.521  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                       bitcoin.json  thrpt    5  28632.668 ±  145.738  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                      doj-blog.json  thrpt    5   4942.639 ±  278.694  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel              eu-lobby-country.json  thrpt    5  35789.824 ±  656.062  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel            eu-lobby-financial.json  thrpt    5   6559.268 ±  283.716  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                 eu-lobby-repr.json  thrpt    5   2618.336 ±   45.202  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                 github-events.json  thrpt    5   3231.984 ±  519.013  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  github-gists.json  thrpt    5   6897.294 ±  388.148  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                json-generator.json  thrpt    5  35760.721 ± 1512.537  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                    meteorites.json  thrpt    5   1213.303 ±    1.975  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                        movies.json  thrpt    5    101.052 ±    0.990  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  reddit-scala.json  thrpt    5   2364.257 ±   16.566  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                    rick-morty.json  thrpt    5  25945.617 ±  107.813  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  temp-anomaly.json  thrpt    5  45026.964 ± 1042.502  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  thai-cinemas.json  thrpt    5  34961.122 ±  432.483  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                       turkish.json  thrpt    5    547.488 ±    5.704  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel  twitter_api_compact_response.json  thrpt    5  44233.710 ± 1428.423  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel          twitter_api_response.json  thrpt    5  34833.882 ± 1169.178  ops/s
```
### GraalVM CE 19.1.0
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                                       (fileName)   Mode  Cnt      Score     Error  Units
[info] JsoniterScalaModelBenchmark.decodeModel                 australia-abc.json  thrpt    5  18353.567 ± 354.867  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                       bitcoin.json  thrpt    5  26931.625 ± 264.183  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                      doj-blog.json  thrpt    5   3347.842 ±  19.738  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel              eu-lobby-country.json  thrpt    5  40682.642 ± 619.812  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel            eu-lobby-financial.json  thrpt    5   6045.554 ±  26.534  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                 eu-lobby-repr.json  thrpt    5   2471.593 ±  25.471  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                 github-events.json  thrpt    5   3232.488 ± 169.090  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  github-gists.json  thrpt    5   8398.183 ± 737.105  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                json-generator.json  thrpt    5  36862.318 ± 417.467  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                    meteorites.json  thrpt    5   1296.611 ±  22.989  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                        movies.json  thrpt    5     86.979 ±   1.645  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  reddit-scala.json  thrpt    5   3831.488 ±   5.594  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                    rick-morty.json  thrpt    5  27888.082 ± 360.005  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  temp-anomaly.json  thrpt    5  42640.140 ± 731.200  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  thai-cinemas.json  thrpt    5  38442.599 ± 324.700  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                       turkish.json  thrpt    5    543.141 ±   1.307  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel  twitter_api_compact_response.json  thrpt    5  44019.628 ± 309.069  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel          twitter_api_response.json  thrpt    5  33761.713 ± 149.848  ops/s
```
### GraalVM EE 19.1.0
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                                                       (fileName)   Mode  Cnt      Score      Error  Units
[info] JsoniterScalaModelBenchmark.decodeModel                 australia-abc.json  thrpt    5  19080.909 ±  900.436  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                       bitcoin.json  thrpt    5  28323.890 ± 1235.923  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                      doj-blog.json  thrpt    5   3598.308 ±   49.463  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel              eu-lobby-country.json  thrpt    5  40349.295 ±  759.972  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel            eu-lobby-financial.json  thrpt    5   7119.765 ±   71.175  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                 eu-lobby-repr.json  thrpt    5   2704.805 ±   28.476  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                 github-events.json  thrpt    5   3607.641 ±   22.937  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  github-gists.json  thrpt    5   8384.975 ±  227.058  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                json-generator.json  thrpt    5  39538.903 ±  134.771  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                    meteorites.json  thrpt    5   1274.925 ±   28.681  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                        movies.json  thrpt    5     93.505 ±    3.823  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  reddit-scala.json  thrpt    5   4270.330 ±   22.915  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                    rick-morty.json  thrpt    5  28000.982 ±  445.126  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  temp-anomaly.json  thrpt    5  44732.399 ±  833.363  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                  thai-cinemas.json  thrpt    5  39780.476 ±  230.824  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel                       turkish.json  thrpt    5    590.186 ±    0.345  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel  twitter_api_compact_response.json  thrpt    5  47179.745 ±  597.709  ops/s
[info] JsoniterScalaModelBenchmark.decodeModel          twitter_api_response.json  thrpt    5  35322.935 ±  308.824  ops/s
```